+++ library/include/hiptensor/hiptensor_types.h
@@ -26,6 +26,7 @@
 #pragma once
 
 #include <stdint.h>
+
 #include <hip/hip_common.h>
 #include <hip/library_types.h>