+++ library/include/hiptensor/internal/config.hpp
@@ -35,6 +35,8 @@ namespace hiptensor
 /// HIPTENSOR_ARCH_GFX90a
 /// HIPTENSOR_ARCH_GFX942
 /// HIPTENSOR_ARCH_GFX950
+/// HIPTENSOR_ARCH_GFX1200
+/// HIPTENSOR_ARCH_GFX1201
 #if defined(__gfx908__)
 #define HIPTENSOR_ARCH_GFX908 __gfx908__
 #elif defined(__gfx90a__)
@@ -43,6 +45,10 @@ namespace hiptensor
 #define HIPTENSOR_ARCH_GFX942 __gfx942__
 #elif defined(__gfx950__)
 #define HIPTENSOR_ARCH_GFX950 __gfx950__
+#elif defined(__gfx1200__)
+#define HIPTENSOR_ARCH_GFX1200 __gfx1200__
+#elif defined(__gfx1201__)
+#define HIPTENSOR_ARCH_GFX1201 __gfx1201__
 #else
 #define HIPTENSOR_ARCH_HOST 1
 #endif
@@ -58,6 +64,12 @@ namespace hiptensor
 #endif
 #if !defined(HIPTENSOR_ARCH_GFX950)
 #define HIPTENSOR_ARCH_GFX950 0
+#endif
+#if !defined(HIPTENSOR_ARCH_GFX1200)
+#define HIPTENSOR_ARCH_GFX1200 0
+#endif
+#if !defined(HIPTENSOR_ARCH_GFX1201)
+#define HIPTENSOR_ARCH_GFX1201 0
 #endif
 #if !defined(HIPTENSOR_ARCH_HOST)
 #define HIPTENSOR_ARCH_HOST 0