+++ library/src/include/data_types_impl.hpp
@@ -26,8 +26,9 @@
 
 #pragma once
 
-#include "data_types.hpp"
 #include <iostream>
+
+#include "data_types.hpp"
 
 namespace hiptensor
 {
@@ -221,4 +222,3 @@ namespace hiptensor
     // @endcond
 
 } // namespace hiptensor
-