+++ samples/01_contraction/simple_bilinear_contraction_f64_f64_f64_f64_compute_f64.cpp
@@ -28,9 +28,9 @@
 int main(int argc, char* argv[])
 {
     /***************************************
-   * Check device support                 *
-   **************************************/
-    if(!isF32Supported())
+    * Check device support                 *
+    **************************************/
+    if(!isF64Supported() || !isF64MatrixCoreSupported())
     {
         std::cout << "unsupported host device" << std::endl;
         exit(EXIT_FAILURE);