+++ samples/01_contraction/simple_scale_contraction_f32_f32_f32_compute_bf16.cpp
@@ -29,9 +29,9 @@
 int main(int argc, char* argv[])
 {
     /***************************************
-   * Check device support                 *
-   **************************************/
-    if(!isF32Supported())
+    * Check device support                 *
+    **************************************/
+    if(!isF32Supported() || !isF16MatrixCoreSupported())
     {
         std::cout << "unsupported host device" << std::endl;
         exit(EXIT_FAILURE);