+++ test/02_elementwise/elementwise_test.cpp
@@ -58,7 +58,8 @@ namespace hiptensor
     // False = skip test
     bool PermutationTest::checkDevice(hiptensorDataType_t datatype) const
     {
-        return isF32Supported() && ((datatype == HIPTENSOR_R_32F) || (datatype == HIPTENSOR_R_16F));
+        return (isF16Supported() && (datatype == HIPTENSOR_R_16F))
+               || (isF32Supported() && (datatype == HIPTENSOR_R_32F));
     }
 
     bool PermutationTest::checkSizes() const