+++ test/02_elementwise/elementwise_trinary_op_test.cpp
@@ -58,8 +58,8 @@ namespace hiptensor
     // False = skip test
     bool ElementwiseTrinaryOpTest::checkDevice(hiptensorDataType_t datatype) const
     {
-        return (isF32Supported()
-                && ((datatype == HIPTENSOR_R_32F) || (datatype == HIPTENSOR_R_16F)))
+        return (isF16Supported() && (datatype == HIPTENSOR_R_16F))
+               || (isF32Supported() && (datatype == HIPTENSOR_R_32F))
                || (isF64Supported() && (datatype == HIPTENSOR_R_64F));
     }