+++ test/utils.hpp
@@ -112,6 +112,24 @@ inline double getEpsilon(hiptensorComputeDescriptor_t id)
     }
 }
 
+inline bool isF16Supported()
+{
+    hipDevice_t     mHandle;
+    hipDeviceProp_t mProps;
+
+    CHECK_HIP_ERROR(hipGetDevice(&mHandle));
+    CHECK_HIP_ERROR(hipGetDeviceProperties(&mProps, mHandle));
+
+    std::string deviceName(mProps.gcnArchName);
+
+    return (deviceName.find("gfx908") != std::string::npos)
+           || (deviceName.find("gfx90a") != std::string::npos)
+           || (deviceName.find("gfx942") != std::string::npos)
+           || (deviceName.find("gfx950") != std::string::npos)
+           || (deviceName.find("gfx1200") != std::string::npos)
+           || (deviceName.find("gfx1201") != std::string::npos);
+}
+
 inline bool isF32Supported()
 {
     hipDevice_t     mHandle;
@@ -125,7 +143,9 @@ inline bool isF32Supported()
     return (deviceName.find("gfx908") != std::string::npos)
            || (deviceName.find("gfx90a") != std::string::npos)
            || (deviceName.find("gfx942") != std::string::npos)
-           || (deviceName.find("gfx950") != std::string::npos);
+           || (deviceName.find("gfx950") != std::string::npos)
+           || (deviceName.find("gfx1200") != std::string::npos)
+           || (deviceName.find("gfx1201") != std::string::npos);
 }
 
 inline bool isF64Supported()
@@ -141,6 +161,77 @@ inline bool isF64Supported()
     return (deviceName.find("gfx90a") != std::string::npos)
            || (deviceName.find("gfx942") != std::string::npos)
            || (deviceName.find("gfx950") != std::string::npos);
+}
+
+inline bool isF16MatrixCoreSupported()
+{
+    hipDevice_t     mHandle;
+    hipDeviceProp_t mProps;
+
+    CHECK_HIP_ERROR(hipGetDevice(&mHandle));
+    CHECK_HIP_ERROR(hipGetDeviceProperties(&mProps, mHandle));
+
+    std::string deviceName(mProps.gcnArchName);
+
+    return (deviceName.find("gfx908") != std::string::npos)
+           || (deviceName.find("gfx90a") != std::string::npos)
+           || (deviceName.find("gfx942") != std::string::npos)
+           || (deviceName.find("gfx950") != std::string::npos)
+           || (deviceName.find("gfx1200") != std::string::npos)
+           || (deviceName.find("gfx1201") != std::string::npos);
+}
+
+inline bool isF32MatrixCoreSupported()
+{
+    hipDevice_t     mHandle;
+    hipDeviceProp_t mProps;
+
+    CHECK_HIP_ERROR(hipGetDevice(&mHandle));
+    CHECK_HIP_ERROR(hipGetDeviceProperties(&mProps, mHandle));
+
+    std::string deviceName(mProps.gcnArchName);
+
+    return (deviceName.find("gfx908") != std::string::npos)
+           || (deviceName.find("gfx90a") != std::string::npos)
+           || (deviceName.find("gfx942") != std::string::npos)
+           || (deviceName.find("gfx950") != std::string::npos);
+}
+
+inline bool isF64MatrixCoreSupported()
+{
+    hipDevice_t     mHandle;
+    hipDeviceProp_t mProps;
+
+    CHECK_HIP_ERROR(hipGetDevice(&mHandle));
+    CHECK_HIP_ERROR(hipGetDeviceProperties(&mProps, mHandle));
+
+    std::string deviceName(mProps.gcnArchName);
+
+    return (deviceName.find("gfx90a") != std::string::npos)
+           || (deviceName.find("gfx942") != std::string::npos)
+           || (deviceName.find("gfx950") != std::string::npos);
+}
+
+inline bool isDataType16Bits(hiptensorDataType_t dataType)
+{
+    return (dataType == HIPTENSOR_R_16F || dataType == HIPTENSOR_R_16BF
+            || dataType == HIPTENSOR_C_16F || dataType == HIPTENSOR_C_16BF);
+}
+
+inline bool isDataType32Bits(hiptensorDataType_t dataType)
+{
+    return (dataType == HIPTENSOR_R_32F || dataType == HIPTENSOR_C_32F);
+}
+
+inline bool isDataType64Bits(hiptensorDataType_t dataType)
+{
+    return (dataType == HIPTENSOR_R_64F || dataType == HIPTENSOR_C_64F);
+}
+
+inline bool isDataTypeComplex(hiptensorDataType_t dataType)
+{
+    return (dataType == HIPTENSOR_C_32F || dataType == HIPTENSOR_C_64F
+            || dataType == HIPTENSOR_C_16F || dataType == HIPTENSOR_C_16BF);
 }
 
 template <typename intT1,