+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 Full documentation for hipTensor is available at [rocm.docs.amd.com/projects/hiptensor](https://rocm.docs.amd.com/projects/hipTensor/en/latest/index.html).
 
-## (Unreleased) hipTensor 2.2.0 for ROCm 7.2.0
+## (Unreleased) hipTensor 2.3.0
+
+### Added
+
+* Added support for new GPU target gfx1200.
+* Added support for new GPU target gfx1201.
+
+## hipTensor 2.2.0 for ROCm 7.2.0
 
 ### Added